[Tests] Move PasswordHashTest and correct related namespaces
