+++ tests/phpunit/unit/AccessControl/TokenTokenTest.php
@@ -1,9 +1,10 @@
 <?php
 
-namespace Bolt\Tests;
+namespace Bolt\Tests\AccessControl;
 
 use Bolt\AccessControl\Token\Token;
 use Bolt\Storage\Entity;
+use Bolt\Tests\BoltUnitTest;
 
 /**
  * Test for AccessControl\Token\Token
